updated README
Added more gifs to show different device sizes and orientations, the app icon, and the launch screen. With the personalized launch screen, I noticed that once the app had been started, it wouldn't come up again as clearly if I didn't catch the recording the first time it came up.